+++ raiden_contracts/tests/test_contract_limits.py
@@ -14,6 +14,8 @@ def test_register_three_but_not_four(
         get_token_network_registry,
         secret_registry_contract,
         custom_token_factory,
+        channel_participant_deposit_limit,
+        token_network_deposit_limit,
 ):
     """ Check that TokenNetworkRegistry observes the max number of tokens """
     token_network_registry = get_token_network_registry([
@@ -28,11 +30,27 @@ def test_register_three_but_not_four(
     token1 = custom_token_factory()
     token2 = custom_token_factory()
     token3 = custom_token_factory()
-    token_network_registry.functions.createERC20TokenNetwork(token0.address).transact()
-    token_network_registry.functions.createERC20TokenNetwork(token1.address).transact()
-    token_network_registry.functions.createERC20TokenNetwork(token2.address).transact()
+    token_network_registry.functions.createERC20TokenNetwork(
+        token0.address,
+        channel_participant_deposit_limit,
+        token_network_deposit_limit,
+    ).transact()
+    token_network_registry.functions.createERC20TokenNetwork(
+        token1.address,
+        channel_participant_deposit_limit,
+        token_network_deposit_limit,
+    ).transact()
+    token_network_registry.functions.createERC20TokenNetwork(
+        token2.address,
+        channel_participant_deposit_limit,
+        token_network_deposit_limit,
+    ).transact()
     with pytest.raises(TransactionFailed):
-        token_network_registry.functions.createERC20TokenNetwork(token3.address).transact()
+        token_network_registry.functions.createERC20TokenNetwork(
+            token3.address,
+            channel_participant_deposit_limit,
+            token_network_deposit_limit,
+        ).transact()
 
 
 def test_channel_participant_deposit_limit_value(token_network):